requirements.txt
django>=4.2.7
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
numpy>=1.26.0
opencv-python-headless>=4.8.0
Pillow>=10.1.0
tensorflow-cpu>=2.16.1
gunicorn>=21.2.0
whitenoise>=6.6.0
requests>=2.31.0
python-dotenv>=1.0.0
google-genai>=0.1.0
openai>=1.12.0
    